Fix transfer fees
