+++ packages/web/integrations/axelar/source-chain-config.ts
@@ -17,7 +17,7 @@ export const SourceChainConfigs: {
         ? "0x254d06f33bDc5b8ee05b2ea472107E300226659A"
         : "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48", // test: 'aUSDC' on metamask/etherscan
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: IS_TESTNET ? "150000" : "10500000", // From https://docs.axelar.dev/resources/mainnet#cross-chain-relayer-gas-fee
+      transferFeeMinAmount: IS_TESTNET ? "150000" : "20500000", // From https://docs.axelar.dev/resources/mainnet#cross-chain-relayer-gas-fee
     },
     avalanche: {
       id: "Avalanche" as const,
@@ -73,7 +73,7 @@ export const SourceChainConfigs: {
         ? "0xB4FBF271143F4FBf7B91A5ded31805e42b2208d6"
         : "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: IS_TESTNET ? "100000000000000" : "6300000000000000",
+      transferFeeMinAmount: IS_TESTNET ? "100000000000000" : "20400000000000000",
     },
   },
   wglmr: {
@@ -87,7 +87,7 @@ export const SourceChainConfigs: {
       logoUrl: "/networks/moonbeam.svg",
       transferFeeMinAmount: IS_TESTNET
         ? "40000000000000000"
-        : "2000000000000000000",
+        : "2900000000000000000",
     },
   },
   wbtc: {
@@ -103,7 +103,7 @@ export const SourceChainConfigs: {
       id: "Ethereum" as const,
       erc20ContractAddress: "0x6B175474E89094C44Da98b954EedeAC495271d0F",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: "10000000000000000000",
+      transferFeeMinAmount: "20500000000000000000",
     },
   },
   dot: {
@@ -111,7 +111,7 @@ export const SourceChainConfigs: {
       id: "Moonbeam" as const,
       erc20ContractAddress: "0xFfFFfFff1FcaCBd218EDc0EbA20Fc2308C778080",
       logoUrl: "/networks/moonbeam.svg",
-      transferFeeMinAmount: "1000000000",
+      transferFeeMinAmount: "1700000000",
     },
   },
   usdt: {
@@ -119,7 +119,7 @@ export const SourceChainConfigs: {
       id: "Ethereum" as const,
       erc20ContractAddress: "0xdAC17F958D2ee523a2206206994597C13D831ec7",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: "10000000",
+      transferFeeMinAmount: "20500000000000000000",
     },
   },
   frax: {
@@ -127,7 +127,7 @@ export const SourceChainConfigs: {
       id: "Ethereum" as const,
       erc20ContractAddress: "0x853d955aCEf822Db058eb8505911ED77F175b99e",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: "10000000000000000000",
+      transferFeeMinAmount: "20500000000000000000",
     },
   },
   link: {
@@ -135,7 +135,7 @@ export const SourceChainConfigs: {
       id: "Ethereum" as const,
       erc20ContractAddress: "0x514910771AF9Ca656af840dff83E8264EcF986CA",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: "2000000000000000000",
+      transferFeeMinAmount: "3080000000000000000",
     },
   },
   aave: {
@@ -143,7 +143,7 @@ export const SourceChainConfigs: {
       id: "Ethereum" as const,
       erc20ContractAddress: "0x7Fc66500c84A76Ad7e9c93437bFc5Ac33E2DDaE9",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: "200000000000000000",
+      transferFeeMinAmount: "308000000000000000",
     },
   },
   ape: {
@@ -151,7 +151,7 @@ export const SourceChainConfigs: {
       id: "Ethereum" as const,
       erc20ContractAddress: "0x4d224452801ACEd8B2F0aebE155379bb5D594381",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: "2000000000000000000",
+      transferFeeMinAmount: "4100000000000000000",
     },
   },
   axs: {
@@ -159,7 +159,7 @@ export const SourceChainConfigs: {
       id: "Ethereum" as const,
       erc20ContractAddress: "0xBB0E17EF65F82Ab018d8EDd776e8DD940327B28b",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: "700000000000000000",
+      transferFeeMinAmount: "1030000000000000000",
     },
   },
   mkr: {
@@ -167,7 +167,7 @@ export const SourceChainConfigs: {
       id: "Ethereum" as const,
       erc20ContractAddress: "0x9f8F72aA9304c8B593d555F12eF6589cC3A579A2",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: "10000000000000000",
+      transferFeeMinAmount: "20600000000000000",
     },
   },
   rai: {
@@ -175,7 +175,7 @@ export const SourceChainConfigs: {
       id: "Ethereum" as const,
       erc20ContractAddress: "0x03ab458634910AaD20eF5f1C8ee96F1D6ac54919",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: "3000000000000000000",
+      transferFeeMinAmount: "7200000000000000000",
     },
   },
   shib: {
@@ -183,7 +183,7 @@ export const SourceChainConfigs: {
       id: "Ethereum" as const,
       erc20ContractAddress: "0x95aD61b0a150d79219dCF64E1E6Cc01f0B64C4cE",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: "1000000000000000000000000",
+      transferFeeMinAmount: "2050000000000000000000000",
     },
   },
   uni: {
@@ -191,7 +191,7 @@ export const SourceChainConfigs: {
       id: "Ethereum" as const,
       erc20ContractAddress: "0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: "2000000000000000000",
+      transferFeeMinAmount: "4100000000000000000",
     },
   },
   xcn: {
@@ -199,7 +199,7 @@ export const SourceChainConfigs: {
       id: "Ethereum" as const,
       erc20ContractAddress: "0xA2cd3D43c775978A96BdBf12d733D5A1ED94fb18",
       logoUrl: "/networks/ethereum.svg",
-      transferFeeMinAmount: "100000000000000000000",
+      transferFeeMinAmount: "206000000000000000000",
     },
   },
   wbnb: {
